+++ src/models/__tests__/Project.spec.ts
@@ -50,16 +50,16 @@ describe('ProjectSearchResult', () => {
         { name: 'foo@example.com', accessLevel: 'Comment', orcid: '0000-0000-0000-1234' },
         { name: 'Jane Doe', accessLevel: 'Own', orcid: '0000-0000-0000-5678' }
       ],
-      contributorsData: 'John Smith|Principal Investigator (PI)|0000-0000-0000-TEST,' +
+      membersData: 'John Smith|Principal Investigator (PI)|0000-0000-0000-TEST,' +
                           'John Smith|Other|0000-0000-0000-TEST,Elmer Fudd|Other|0000-0000-0000-9876',
-      contributors: [
+      members: [
         { name: 'John Smith', role: 'Principal Investigator (PI), Other', orcid: '0000-0000-0000-TEST' },
         { name: 'Elmer Fudd', role: 'Other', orcid: '0000-0000-0000-9876' }
       ],
-      fundersData: 'Test Funder|12345,Another Funder|67890',
-      funders: [
-        { name: 'Test Funder', grantId: '12345' },
-        { name: 'Another Funder', grantId: '67890' },
+      fundingsData: 'Test funding|12345,Another funding|67890',
+      fundings: [
+        { name: 'Test funding', grantId: '12345' },
+        { name: 'Another funding', grantId: '67890' },
       ],
     });
   });
@@ -86,9 +86,9 @@ describe('ProjectSearchResult', () => {
         modified: projectSearchResult.modified,
         modifiedByName: projectSearchResult.modifiedByName,
         collaboratorsData: 'foo@example.com|Comment|0000-0000-0000-1234,Jane Doe|Own|0000-0000-0000-5678',
-        contributorsData: 'John Smith|Principal Investigator (PI)|0000-0000-0000-TEST,' +
+        membersData: 'John Smith|Principal Investigator (PI)|0000-0000-0000-TEST,' +
                           'John Smith|Other|0000-0000-0000-TEST,Elmer Fudd|Other|0000-0000-0000-9876',
-        fundersData: 'Test Funder|12345,Another Funder|67890',
+        fundingsData: 'Test funding|12345,Another funding|67890',
       }
       localQuery.mockResolvedValueOnce({ items: [queryResult] });
 
@@ -114,20 +114,20 @@ describe('ProjectSearchResult', () => {
                               'END, ' +
                               'r.label, ' +
                               'pc.orcid ' +
-                          ') ORDER BY pc.created) as contributorsData, ' +
-                          'GROUP_CONCAT(DISTINCT CONCAT_WS(\'|\', funders.name, pf.grantId) ' +
-                            'ORDER BY funders.name SEPARATOR \',\') fundersData ' +
+                          ') ORDER BY pc.created) as membersData, ' +
+                          'GROUP_CONCAT(DISTINCT CONCAT_WS(\'|\', fundings.name, pf.grantId) ' +
+                            'ORDER BY fundings.name SEPARATOR \',\') fundingsData ' +
                         'FROM projects p ' +
                           'LEFT JOIN researchDomains ON p.researchDomainId = researchDomains.id ' +
                           'LEFT JOIN users cu ON cu.id = p.createdById ' +
                           'LEFT JOIN users mu ON mu.id = p.modifiedById ' +
                           'LEFT JOIN projectCollaborators pcol ON pcol.projectId = p.id ' +
                             'LEFT JOIN users collab ON pcol.userId = collab.id ' +
-                          'LEFT JOIN projectContributors pc ON pc.projectId = p.id ' +
-                            'LEFT JOIN projectContributorRoles pcr ON pc.id = pcr.projectContributorId ' +
-                            'LEFT JOIN contributorRoles r ON pcr.contributorRoleId = r.id ' +
-                          'LEFT JOIN projectFunders pf ON pf.projectId = p.id ' +
-                            'LEFT JOIN affiliations funders ON pf.affiliationId = funders.uri ';
+                          'LEFT JOIN projectMembers pc ON pc.projectId = p.id ' +
+                            'LEFT JOIN projectMemberRoles pcr ON pc.id = pcr.projectMemberId ' +
+                            'LEFT JOIN memberRoles r ON pcr.memberRoleId = r.id ' +
+                          'LEFT JOIN projectFundings pf ON pf.projectId = p.id ' +
+                            'LEFT JOIN affiliations fundings ON pf.affiliationId = fundings.uri ';
       const vals = [`%${term.toLowerCase()}%`, `%${term.toLowerCase()}%`,
                     projectSearchResult.createdById.toString(), projectSearchResult.createdById.toString()];
       const whereFilters = ['(LOWER(p.title) LIKE ? OR LOWER(p.abstractText) LIKE ?)',