+++ src/resolvers/__tests__/plan.spec.ts
@@ -32,8 +32,8 @@ import { clearResearchDomainStore, initResearchDomainStore, mockFindResearchDoma
 import { initUserStore, mockFindUserById } from "../../models/__mocks__/User";
 import { RelatedWork } from "../../models/RelatedWork";
 import { clearRelatedWorkStore, initRelatedWorkStore, mockFindRelatedWorksByProjectId } from "../../models/__mocks__/RelatedWork";
-import { ContributorRole } from "../../models/ContributorRole";
-import { clearContributorRoles, initContributorRoles, mockDefaultContributorRole } from "../../models/__mocks__/ContributorRole";
+import { MemberRole } from "../../models/MemberRole";
+import { clearMemberRoles, initMemberRoles, mockDefaultMemberRole } from "../../models/__mocks__/MemberRole";
 import { ProjectCollaborator, ProjectCollaboratorAccessLevel } from "../../models/Collaborator";
 import {
   clearProjectCollaboratorsStore,
@@ -42,8 +42,8 @@ import {
 } from "../../models/__mocks__/Collaborator";
 import { DynamoDBClient, PutItemCommandInput, QueryCommandInput } from "@aws-sdk/client-dynamodb";
 import { awsConfig } from "../../config/awsConfig";
-import { PlanFunder } from "../../models/Funder";
-import { PlanContributor } from "../../models/Contributor";
+import { PlanFunding } from "../../models/Funding";
+import { PlanMember } from "../../models/Member";
 import { Answer } from "../../models/Answer";
 import { generalConfig } from "../../config/generalConfig";
 import { getCurrentDate } from "../../utils/helpers";
@@ -56,7 +56,7 @@ let testServer: ApolloServer;
 
 let userStore: User[];
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-let contributorRoleStore: ContributorRole[];
+let memberRoleStore: MemberRole[];
 let projectCollaboratorStore: ProjectCollaborator[];
 let researchStore: ResearchDomain[];
 let relatedWorkStore: RelatedWork[];
@@ -122,7 +122,7 @@ beforeEach(async () => {
   versionedTemplateStore = initVersionedTemplateStore(1);
   projectStore = initProjectStore(1);
   projectCollaboratorStore = initProjectCollaboratorsStore(1);
-  contributorRoleStore = initContributorRoles(1);
+  memberRoleStore = initMemberRoles(1);
   relatedWorkStore = initRelatedWorkStore(1);
   researchStore = initResearchDomainStore(1);
   planStore = initPlanStore(3);
@@ -168,11 +168,11 @@ beforeEach(async () => {
   jest.spyOn(VersionedTemplate, 'findById').mockImplementation(mockFindVersionedTemplateById);
   jest.spyOn(ResearchDomain, 'findById').mockImplementation(mockFindResearchDomainById);
   jest.spyOn(RelatedWork, 'findByProjectId').mockImplementation(mockFindRelatedWorksByProjectId);
-  jest.spyOn(ContributorRole, 'defaultRole').mockImplementation(mockDefaultContributorRole);
+  jest.spyOn(MemberRole, 'defaultRole').mockImplementation(mockDefaultMemberRole);
 
   // Mock the commonStandardService queries
-  jest.spyOn(PlanFunder, 'query').mockResolvedValue([]);
-  jest.spyOn(PlanContributor, 'query').mockResolvedValue([]);
+  jest.spyOn(PlanFunding, 'query').mockResolvedValue([]);
+  jest.spyOn(PlanMember, 'query').mockResolvedValue([]);
   jest.spyOn(User, 'query').mockResolvedValue([userStore[0]]);
   jest.spyOn(Answer, 'query').mockResolvedValue([]);
 
@@ -193,7 +193,7 @@ afterEach(() => {
   clearProjectCollaboratorsStore();
   clearPlanStore();
   clearPlanVersionStore();
-  clearContributorRoles();
+  clearMemberRoles();
   clearResearchDomainStore();
   clearRelatedWorkStore();
 });
@@ -214,8 +214,8 @@ describe('plans query', () => {
           dmpId
           registeredBy
           registered
-          funder
-          contributors
+          funding
+          members
           templateTitle
         }
       }
@@ -237,8 +237,8 @@ describe('plans query', () => {
     expect(resp.body.singleResult.data?.plans[0]?.registered).toEqual(planStore[0].registered);
     expect(resp.body.singleResult.data?.plans[0]?.dmpId).toEqual(planStore[0].dmpId);
     expect(resp.body.singleResult.data?.plans[0]?.registeredBy).toBeTruthy();
-    expect(resp.body.singleResult.data?.plans[0]?.funder).toBeTruthy();
-    expect(resp.body.singleResult.data?.plans[0]?.contributors).toBeTruthy();
+    expect(resp.body.singleResult.data?.plans[0]?.funding).toBeTruthy();
+    expect(resp.body.singleResult.data?.plans[0]?.members).toBeTruthy();
     expect(resp.body.singleResult.data?.plans[0]?.templateTitle).toBeTruthy();
 
     expect(resp.body.singleResult.data?.plans[0]?.created).toEqual(planStore[0].created);