+++ CHANGELOG.md
@@ -94,6 +94,7 @@
 ### Updated
 - Updated the `Question` schema to allow `questionJSON: String` to be used in GraphQL query
 - Updated `Question` model to access `questionJSON` input, parse the JSON using the Zod schemas provided by `@dmptool/types`, and populate the `questionType` with the parsed JSON.
+- Rename all occurrences of `Contributor` to `Member` and `Funder` to `Funding` to match newer terminology
 - Updated `addSection` resolver to properly copy questions and tags when creating a new section by copying one
 - Replaced all instances of `TemplateVisibility.PRIVATE` with `TemplateVisibility.ORGANIZATION` [#159]
 - Renamed the `outputTypes` table to `projectOutputTypes`